Use extension of the template file when creating a migration
Hi! This is a very nice tool to use. Though there is one inconvenience with it. It doesn't support transpiled languages like TypeScript.
It would be great if there was an option like `--createDir` to specify where the template file goes when creating a migration, because otherwise it goes to the `build` dir with compiled JavaScript files...
Anyway this PR solves a small inconvenience where my adapter for Elasticsearch specifies a path to `template.ts`, but when I run
`east create my-migration` it is created as `1_my-migration.js` instead, which is annoying.